requirements.txt
google-genai>=1.0.0
mss>=9.0.0
Pillow>=10.0.0
